feat(sl-types)!: widen GridRectangle to u32 and migrate more value types from sl-client
Widen the grid-index coordinate family to u32 so the Second Life
whole-grid map layer's rectangle bounds (which can exceed u16::MAX) can
be represented without truncation:

- GridCoordinates / GridCoordinateOffset / GridRectangle and the
  GridRectangleLike trait (size_x/size_y, x_range/y_range, intersect)
  now work in u32; the Add/Sub coordinate arithmetic widens its
  intermediates to i64 and saturates.
- pps_hud_config is re-expressed through the new
  GlobalCoordinates::from_grid_corner.
- Consumers (sl-glw, sl-map-apis, sl-map-cli, sl-map-web) are updated
  for the widening; grid-coordinate-bearing DTO / request / CLI-arg
  fields widen to u32 (serde and clap parse u32 unchanged).
- The redb region-name caches move to *_v2 tables (their grid-coordinate
  components widen u16 -> u32); the legacy u16 tables are dropped on open
  to reclaim space (pure caches, regenerated from the grid).

Migrate the remaining general value types created client-local in
sl-client (per the "new types go local first, batch-migrate later"
rule) into the shared crate:

- map.rs: Direction, GlobalCoordinates, Scale, TeleportFlags
- key.rs: PickKey, GroupNoticeKey (wrapping Key like the other keys)
- lsl.rs: ScriptPermissions
- new environment.rs: Color, ColorAlpha, Glow, CloudPosDensity
- new friend.rs: FriendRights
- new experience.rs: ExperienceProperties (+ PROPERTY_* consts)
- new parcel.rs: ParcelAccessFlags, ParcelReturnType


Ran_pre-commit_check: frontend:::biome-format
Ran_pre-commit_check: frontend:::biome-lint
Ran_pre-commit_check: lsl:::lslint
Ran_pre-commit_check: markdown:::code-blocks:::shellcheck-bash
Ran_pre-commit_check: markdown:::code-blocks:::shellcheck-sh
Ran_pre-commit_check: markdown:::code-blocks:::shfmt
Ran_pre-commit_check: markdown:::rumdl
Ran_pre-commit_check: nodejs:::prettier
Ran_pre-commit_check: shell:::shellcheck
Ran_pre-commit_check: shell:::shfmt
Ran_pre-commit_check: sql:::sqlfluff
Ran_pre-commit_check: sql:::sqlfluff-format
Ran_pre-commit_check: toml:::tombi
Ran_pre-commit_check: toml:::tombi-fmt
Ran_pre-commit_check: common:::whitespace
Ran_pre-commit_check: common:::non-ascii-filenames
Ran_pre-commit_check: common:::typos
Ran_pre-commit_check: github:::pinact-check
Ran_pre-commit_check: github:::wrkflw-validate
Ran_pre-commit_check: github:::zizmor
Ran_pre-commit_check: rust:::bin:::debian-changelog-includes-version
Ran_pre-commit_check: rust:::cargo-check
Ran_pre-commit_check: rust:::cargo-deny-advisories
Ran_pre-commit_check: rust:::cargo-deny-licenses
Ran_pre-commit_check: rust:::cargo-deny-sources
Ran_pre-commit_check: rust:::cargo-deny-workspace-member-deny-toml
Ran_pre-commit_check: rust:::cargo-hack-feature-powerset
Ran_pre-commit_check: rust:::cargo-lock-committed
Ran_pre-commit_check: rust:::cargo-sort
Ran_pre-commit_check: rust:::changelog-md-includes-version
Ran_pre-commit_check: rust:::clippy
Ran_pre-commit_check: rust:::crate-version-not-in-cargo-lock
Ran_pre-commit_check: rust:::doc
Ran_pre-commit_check: rust:::formatting
Ran_pre-commit_check: rust:::nextest
Ran_pre-commit_check: rust:::sqlx-prepare-data-outdated
Ran_pre-commit_check: rust:::unused-dependencies
Ran_pre-commit_check: systemd:::analyze-verify
Skipped_pre-commit_check: automatic;;;rust:::lib:::semver-violations;;;This is not the commit that increments the version number yet
Skipped_pre-commit_check: automatic;;;rust:::lib:::semver-violations;;;This is not the commit that increments the version number yet
Skipped_pre-commit_check: automatic;;;rust:::lib:::semver-violations;;;This is not the commit that increments the version number yet
Skipped_pre-commit_check: automatic;;;rust:::lib:::semver-violations;;;This is not the commit that increments the version number yet
Skipped_pre-commit_check: automatic;;;rust:::lib:::semver-violations;;;This is not the commit that increments the version number yet

+++ sl-glw/examples/render_sample.rs
@@ -153,7 +153,7 @@ fn main() -> Result<(), Box<dyn std::error::Error>> {
         GridCoordinates::new(1140, 1055),
     );
     let zoom = ZoomLevel::try_new(2)?;
-    let image_size: u32 = u32::from(grid.size_x()) * u32::from(zoom.pixels_per_region());
+    let image_size: u32 = grid.size_x() * u32::from(zoom.pixels_per_region());
 
     // Mid-blue background so the white arrows / green outlines stand
     // out without needing real map tiles for the demo.
